Implement player moving

+++ pkg/model/game.go
@@ -1,6 +1,6 @@
 package model
 
 type Position struct {
-	X int `json:"x"`
-	Y int `json:"y"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
 }
